Updated code for 1-click installation

+++ DDLs/SCH-Job-[(dba) Remove-XEventFiles].sql
@@ -70,5 +70,6 @@ EndSave:
 GO
 
 
-EXEC msdb.dbo.sp_start_job @job_name=N'(dba) Remove-XEventFiles'
+IF APP_NAME() = 'Microsoft SQL Server Management Studio - Query'
+	EXEC msdb.dbo.sp_start_job @job_name=N'(dba) Remove-XEventFiles'
 GO
